Separate out the HTML code from jupybakeutils
